Add common transactions steps
